+++ include/ESP32TapeCartLoader.h
@@ -10,10 +10,11 @@ namespace TapuinoNext
     class ESP32TapeCartLoader
     {
       public:
-        ESP32TapeCartLoader(UtilityCollection* utilityCollection);
+        ESP32TapeCartLoader(UtilityCollection* utilityCollection, File& prgFile);
         ~ESP32TapeCartLoader();
         void Init();
         bool CheckForMode();
+        bool LoadPRG();
 
       protected:
         void MotorSignalCallback(bool writeHigh);
@@ -23,10 +24,12 @@ namespace TapuinoNext
       private:
         void FastSendByte(uint8_t byte);
         UtilityCollection* utilityCollection;
+        FlipBuffer* flipBuffer;
         static ESP32TapeCartLoader* internalClass;
         static void IRAM_ATTR MotorSignalCallbackStatic();
         volatile uint16_t shiftReg;
         bool loaderMode;
+        File prgFile;
         uint16_t callAddr, loadAddr, endAddr;
     };
 } // namespace TapuinoNext