+++ include/TapBase.h
@@ -78,14 +78,7 @@ namespace TapuinoNext
         InputHandler* inputHandler;
         LCDUtils* lcdUtils;
         Options* options;
-
-        uint8_t* pBuffer;                       // buffer of raw TAP data
-        uint32_t bufferSize;                    // full size of the TAP buffer this is a ^2
-        uint32_t halfBufferSize;                // half size of the TAP buffer
-        uint32_t bufferPos;                     // current position in the TAP buffer
-        uint32_t bufferMask;                    // bit mask used to reset the bufferPos as it moves passed the bufferSize
-        uint32_t bufferSwitchPos;               // position at which to switch to the "next" half of the buffer, alternates between halfBufferSize and 0
-        volatile bool bufferSwitchFlag = false; // flag set by ReadByte that notifies the main loop to fill the next half of the buffer
+        FlipBuffer* flipBuffer;
 
         // cycle multipliers are PAL / NSTC and machine dependent and calculated at run time
         // these values convert the TAP data into the number of microseconds that the signal last for (high and low edge)